+++ src/ImageGalleryServiceProvider.php
@@ -2,9 +2,11 @@
 
 namespace Alsaloul\ImageGallery;
 
+use Filament\Infolists\Components\ImageEntry;
 use Filament\Support\Assets\Css;
 use Filament\Support\Assets\Js;
 use Filament\Support\Facades\FilamentAsset;
+use Filament\Tables\Columns\ImageColumn;
 use Spatie\LaravelPackageTools\Package;
 use Spatie\LaravelPackageTools\PackageServiceProvider;
 
@@ -29,6 +31,21 @@ class ImageGalleryServiceProvider extends PackageServiceProvider
             $this->getAssets(),
             $this->getAssetPackageName()
         );
+
+        ImageColumn::macro('imageGallery', function () {
+            /** @var ImageColumn $this */
+            $this->view('image-gallery::columns.image-column-gallery');
+            $this->disabledClick();
+
+            return $this;
+        });
+
+        ImageEntry::macro('imageGallery', function () {
+            /** @var ImageEntry $this */
+            $this->view('image-gallery::infolists.entries.image-entry-gallery');
+
+            return $this;
+        });
     }
 
     protected function getAssetPackageName(): ?string